# Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation MeshGroup Test Work",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {
      "name": "Pavel SUntsov",
      "email": "p.v.suntsov@gmail.com"
    }
  },
  "host": "localhost:8010",
  "basePath": "/",
  "tags": [
    {
      "name": "error-controller",
      "description": "Контроллер отвечающий за получение ошибок"
    },
    {
      "name": "profile-controller",
      "description": "Контроллер отвечающий за работу с профилями"
    }
  ],
  "paths": {
    "/error/last": {
      "get": {
        "tags": [
          "error-controller"
        ],
        "summary": "Получение последней ошибки",
        "operationId": "getLastErrorUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ErrorDtoOut"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    },
    "/exit": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Остановка приложения (не реализовано)",
        "operationId": "exitUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "204": {
            "description": "No content"
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Получение всех профилей",
        "operationId": "getAllProfilesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProfileDtoOutGet"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    },
    "/profiles/get": {
      "post": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Получение профиля по e-mail",
        "operationId": "getProfileByEmailUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "profileEmail",
            "description": "profileEmail",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileDtoInEmail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutGet"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    },
    "/profiles/last": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Получение последнего созданного профиля",
        "operationId": "getLastUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutGet"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    },
    "/profiles/set": {
      "post": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Создание нового профиля",
        "operationId": "createProfileUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "profileDtoIn",
            "description": "profileDtoIn",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileDtoIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutPost"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    },
    "/profiles/{id}": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "summary": "Получение профиля по E-mail",
        "operationId": "getProfileByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutGet"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProfileDtoOutMsg"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ErrorDtoOut": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        },
        "created": {
          "$ref": "#/definitions/Timestamp"
        }
      },
      "title": "ErrorDtoOut"
    },
    "ModelAndView": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "boolean"
        },
        "model": {
          "type": "object"
        },
        "modelMap": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "reference": {
          "type": "boolean"
        },
        "status": {
          "type": "string",
          "enum": [
            "ACCEPTED",
            "ALREADY_REPORTED",
            "BAD_GATEWAY",
            "BAD_REQUEST",
            "BANDWIDTH_LIMIT_EXCEEDED",
            "CHECKPOINT",
            "CONFLICT",
            "CONTINUE",
            "CREATED",
            "DESTINATION_LOCKED",
            "EXPECTATION_FAILED",
            "FAILED_DEPENDENCY",
            "FORBIDDEN",
            "FOUND",
            "GATEWAY_TIMEOUT",
            "GONE",
            "HTTP_VERSION_NOT_SUPPORTED",
            "IM_USED",
            "INSUFFICIENT_SPACE_ON_RESOURCE",
            "INSUFFICIENT_STORAGE",
            "INTERNAL_SERVER_ERROR",
            "I_AM_A_TEAPOT",
            "LENGTH_REQUIRED",
            "LOCKED",
            "LOOP_DETECTED",
            "METHOD_FAILURE",
            "METHOD_NOT_ALLOWED",
            "MOVED_PERMANENTLY",
            "MOVED_TEMPORARILY",
            "MULTIPLE_CHOICES",
            "MULTI_STATUS",
            "NETWORK_AUTHENTICATION_REQUIRED",
            "NON_AUTHORITATIVE_INFORMATION",
            "NOT_ACCEPTABLE",
            "NOT_EXTENDED",
            "NOT_FOUND",
            "NOT_IMPLEMENTED",
            "NOT_MODIFIED",
            "NO_CONTENT",
            "OK",
            "PARTIAL_CONTENT",
            "PAYLOAD_TOO_LARGE",
            "PAYMENT_REQUIRED",
            "PERMANENT_REDIRECT",
            "PRECONDITION_FAILED",
            "PRECONDITION_REQUIRED",
            "PROCESSING",
            "PROXY_AUTHENTICATION_REQUIRED",
            "REQUESTED_RANGE_NOT_SATISFIABLE",
            "REQUEST_ENTITY_TOO_LARGE",
            "REQUEST_HEADER_FIELDS_TOO_LARGE",
            "REQUEST_TIMEOUT",
            "REQUEST_URI_TOO_LONG",
            "RESET_CONTENT",
            "SEE_OTHER",
            "SERVICE_UNAVAILABLE",
            "SWITCHING_PROTOCOLS",
            "TEMPORARY_REDIRECT",
            "TOO_EARLY",
            "TOO_MANY_REQUESTS",
            "UNAUTHORIZED",
            "UNAVAILABLE_FOR_LEGAL_REASONS",
            "UNPROCESSABLE_ENTITY",
            "UNSUPPORTED_MEDIA_TYPE",
            "UPGRADE_REQUIRED",
            "URI_TOO_LONG",
            "USE_PROXY",
            "VARIANT_ALSO_NEGOTIATES"
          ]
        },
        "view": {
          "$ref": "#/definitions/View"
        },
        "viewName": {
          "type": "string"
        }
      },
      "title": "ModelAndView"
    },
    "ProfileDtoIn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "ProfileDtoIn"
    },
    "ProfileDtoInEmail": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "title": "ProfileDtoInEmail"
    },
    "ProfileDtoOutGet": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "created": {
          "$ref": "#/definitions/Timestamp"
        }
      },
      "title": "ProfileDtoOutGet"
    },
    "ProfileDtoOutMsg": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        }
      },
      "title": "ProfileDtoOutMsg"
    },
    "ProfileDtoOutPost": {
      "type": "object",
      "properties": {
        "idUser": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "ProfileDtoOutPost"
    },
    "Timestamp": {
      "type": "object",
      "properties": {
        "date": {
          "type": "integer",
          "format": "int32"
        },
        "day": {
          "type": "integer",
          "format": "int32"
        },
        "hours": {
          "type": "integer",
          "format": "int32"
        },
        "minutes": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "nanos": {
          "type": "integer",
          "format": "int32"
        },
        "seconds": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "integer",
          "format": "int64"
        },
        "timezoneOffset": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Timestamp"
    },
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}