# src/main/java/ru/sunlab/meshgrouptest/dto/profile/ProfileDtoIn.java
package ru.sunlab.meshgrouptest.dto.profile;

import lombok.Getter;
import lombok.Setter;

import javax.validation.constraints.Email;
import javax.validation.constraints.NotEmpty;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.Positive;

@Getter
@Setter
public class ProfileDtoIn {

    @NotEmpty
    private String name;

    @NotEmpty
    @Email
    private String email;

    @NotNull
    @Positive
    private int age;
}